// apps/web/lib/api/client/models/PatientUpdate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PatientUpdate = {
    patient_fname?: (string | null);
    patient_mname?: (string | null);
    patient_lname?: (string | null);
    patient_dob?: (string | null);
    patient_contact_phone?: (string | null);
    patient_contact_email?: (string | null);
    patient_gender?: (string | null);
    patient_insurance_member_id?: (string | null);
};